CHANGE schemacheck as own module
